Refact models
