+++ challangeMarvel/HTTPClient/HTTPClient.swift
@@ -1,11 +1,3 @@
-//
-//  HTTPClient.swift
-//  challangeMarvel
-//
-//  Created by Larissa Diniz on 28/07/20.
-//  Copyright © 2020 Larissa Diniz. All rights reserved.
-//
-
 import Foundation
 
 typealias DataTaskResult = (Data?, URLResponse?, Error?) -> Void
