--- challangeMarvel/Models/CharacterModels/CharacterDataContainer.swift
@@ -1,17 +1,0 @@
-//
-//  CharacterDataContainer.swift
-//  challangeMarvel
-//
-//  Created by Larissa Diniz on 28/07/20.
-//  Copyright © 2020 Larissa Diniz. All rights reserved.
-//
-
-import Foundation
-
-struct CharacterDataContainer: Codable {
-    let offset: Int?
-    let limit: Int?
-    let total: Int?
-    let count: Int?
-    let results: [Character]?
-}
