--- challangeMarvel/Models/CharacterModels/CharacterDataWrapper.swift
@@ -1,20 +1,0 @@
-//
-//  CharacterDataWrapper.swift
-//  challangeMarvel
-//
-//  Created by Larissa Diniz on 28/07/20.
-//  Copyright © 2020 Larissa Diniz. All rights reserved.
-//
-
-import Foundation
-
-struct CharacterDataWrapper: Codable {
-    let code: Int?
-    let status: String?
-    let copyright: String?
-    let attributionText: String?
-    let attributionHTML: String?
-    let data: CharacterDataContainer?
-    let etag: String?
-}
-
