--- challangeMarvel/Models/CharacterModels/ComicSummary.swift
@@ -1,15 +1,0 @@
-//
-//  ComicSummary.swift
-//  challangeMarvel
-//
-//  Created by Larissa Diniz on 28/07/20.
-//  Copyright © 2020 Larissa Diniz. All rights reserved.
-//
-
-import Foundation
-
-struct ComicSummary: Codable {
-    let resourceURI: String?
-    let name: String?
-}
-
