+++ challangeMarvel/Models/ComicModels/Comic.swift
@@ -1,12 +1,22 @@
-//
-//  Comic.swift
-//  challangeMarvel
-//
-//  Created by Larissa Diniz on 29/07/20.
-//  Copyright © 2020 Larissa Diniz. All rights reserved.
-//
-
 import Foundation
+
+struct ComicDataContainer: Codable {
+    let offset: Int?
+    let limit: Int?
+    let total: Int?
+    let count: Int?
+    let results: [Comic]?
+}
+
+struct ComicDataWrapper: Codable {
+    let code: Int?
+    let status: String?
+    let copyright: String?
+    let attributionText: String?
+    let attributionHTML: String?
+    let data: ComicDataContainer?
+    let etag: String?
+}
 
 struct Comic: Codable {
     let id: Int?
@@ -16,7 +26,7 @@ struct Comic: Codable {
     func getComicImage() -> String {
         guard let urlPath = thumbnail?.path else { return "" }
         guard let urlExt = thumbnail?.ext else { return "" }
-        let urlImage = "\(urlPath)/portrait_xlarge.\(urlExt)"
-        return urlImage
+        
+        return "\(urlPath)/portrait_xlarge.\(urlExt)"
     }
 }
