--- challangeMarvel/Models/ComicModels/ComicDataContainer.swift
@@ -1,17 +1,0 @@
-//
-//  ComicDataContainer.swift
-//  challangeMarvel
-//
-//  Created by Larissa Diniz on 29/07/20.
-//  Copyright © 2020 Larissa Diniz. All rights reserved.
-//
-
-import Foundation
-
-struct ComicDataContainer: Codable {
-    let offset: Int?
-    let limit: Int?
-    let total: Int?
-    let count: Int?
-    let results: [Comic]?
-}
