--- challangeMarvel/Models/ComicModels/ComicDataWrapper.swift
@@ -1,19 +1,0 @@
-//
-//  ComicDataWrapper.swift
-//  challangeMarvel
-//
-//  Created by Larissa Diniz on 29/07/20.
-//  Copyright © 2020 Larissa Diniz. All rights reserved.
-//
-
-import Foundation
-
-struct ComicDataWrapper: Codable {
-    let code: Int?
-    let status: String?
-    let copyright: String?
-    let attributionText: String?
-    let attributionHTML: String?
-    let data: ComicDataContainer?
-    let etag: String?
-}
