+++ challangeMarvel/Models/Image.swift
@@ -1,11 +1,3 @@
-//
-//  Image.swift
-//  challangeMarvel
-//
-//  Created by Larissa Diniz on 28/07/20.
-//  Copyright © 2020 Larissa Diniz. All rights reserved.
-//
-
 import Foundation
 
 struct Image: Codable {
